pyproject.toml
[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project]
name = "version-tagger"
version = "0.1.14"
description = "Add and bump semantic version tags to commits"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [ "dunamai>=1.23.0", "gitpython>=3.1.44", "toml>=0.10.2", "typer>=0.15.1",]
classifiers = [ "Development Status :: 3 - Alpha", "Programming Language :: Python :: 3.10", "Programming Language :: Python :: 3.11", "Programming Language :: Python :: 3.12", "Programming Language :: Python :: 3.13",]
[[project.authors]]
name = "James Blackwell"
email = "33688964+j-blackwell@users.noreply.github.com"

[project.license]
file = "LICENSE"

[project.urls]
Source = "https://github.com/j-blackwell/version-tagger"

[project.scripts]
bump = "version_tagger:app"

[tool.hatch.build]
include = [ "version_tagger.py",]
